[DOC] hierarchy restructured
The structure was:
- handlers
    |- apps
    |- flows
   - objects.py

But objects.py stored all the objects and type hints related to all apps, so, in order to give it a better design, the structure now is:

- apps
    |- objects
        - [app_name_1].py
        - [app_name_2].py
        - [...]
- [app_name_1].py
- [app_name_2].py
- [...].py

With this, importing is shorter, and flows can be done as a function of the class. If there is needed a flow between apps, that could be manage out the directory.

Also, thanks to an idea on Excel and an error made by the linter, apps module tables are now a column on an Excel file, and with this, all tables are just copied and paste. So if in the future is needed, they are all on the modules file