[RESTRUCTURE] Pip library
When it was a project, the structure was ok, but considering it a library, forces to restructure the directory and naming.

The 'project' was called using 'odoo.apps', after researching, I noticed that this wasn't use in other libraries, so I decided to use it replacing the dot ('.') for a lower dash ('_').

Taking advantage of this restructuring, the return values were replaced by a flask.Response, so now it can be used as an API tool for Odoo's RPC API.

Missing to check some import libraries and some tests.