[FEAT] Request & Response
Added Request and Response Objects to ensure that the actions requested to the client are the ones needed.

This surge due to a pylint error where the `create` function had too many values. Also, this gives structure when using in other calls.

Now, the Responses get the function `complete_response` to update the response according to the given result after an action of the RPC API call.